custom adjustments to NTPClient
now should include fraction-information

print no longer works

+++ arduino_edge_node/NTPClientCustom.h
@@ -105,7 +105,7 @@ class NTPClientCustom {
     /**
      * @return time in seconds since Jan. 1, 1970
      */
-    unsigned long getEpochTime() const;
+    double getEpochTime() const;
 
     /**
      * Stops the underlying UDP client
